Return an empty array when there are no TLS specs
This fixes the "immediate" issue of #3: The return value was assumed to be an array without checking,
so we crashed in case the ingress should get processed but the TLS specs were missing.

This was actually flagged by Visual Studio's checking code, albeit with a hard-to-understand error message
about a "missing callable signature".

Fixes #3